+++ crates/bonsaidb-core/src/limits.rs
@@ -1,4 +1,4 @@
 /// The maximum number of results allowed to be returned from `list_executed_transactions`.
-pub const LIST_TRANSACTIONS_MAX_RESULTS: usize = 1000;
+pub const LIST_TRANSACTIONS_MAX_RESULTS: u32 = 1000;
 /// If no `result_limit` is specified, this value is the limit used by default.
-pub const LIST_TRANSACTIONS_DEFAULT_RESULT_COUNT: usize = 100;
+pub const LIST_TRANSACTIONS_DEFAULT_RESULT_COUNT: u32 = 100;
